Change database export structure
Reworked the template module to not bother with replacements, and
instead allow code to just write to streams themselves with a fixed
header and footer.

The music database now emits lazily loaded data for its search index,
which additionally no longer uses a radix tree more testing with
memory shows that our dataset is marginally smaller with just a
binary-search based approach, and with a lot less code required.