Add support for querying file duration
This adds another table to the database mapping music index numbers
to durations.

Durations are exposed in fractional seconds, so should conform with
the majority of the API.

The exporter isn't finished yet as the current solution involves
generating a 5GB+ cache of music files to parse durations, for
which I'm hoping to avoid. The data has been regenerated in the
meantime.