Add support for querying file duration (#14)
* Add support for querying file duration

This adds another table to the database mapping music index numbers
to durations.

Durations are exposed in fractional seconds, so should conform with
the majority of the API.

* Remove old exporter code

A new repository was set up with the (updated) tool migrated
to it. Another PR will merge in the changes requires for it.