Implement advanced music search and new exporter (#20)
* Implement advanced music search and new exporter

This includes all the necessary changes to pull in support for the
new exporter at a static data level. No support for the compression
is included, just the (minor) tweaks to the data structure for
search indices.

Searching is implemented as described in the associated ticket;
an options table can be supplied to FindMusicFiles with a method
key of "substring", "pattern", or "prefix" to control how the
search is performed.

README updated with examples of advanced search, and applied other
clarifications/corrections.

Fixed a nil return from GetMusicFileDuration when the documented
behaviour was to return a zero value.