Add support for compressed data
This implements the Lua side of things; the exporter can generate
compressed tables as a base64-encoded DEFLATEd string.

Removed the CreateHydratedTable function and replaced it with a
single HydrateData that takes either a table (similar to the old
manner) or a string for decompression on first-access.

At the moment, compression isn't enabled - it takes a bit of time
to thaw the data, and other parts of the API will frequently load
the search tree in this patch for the name => ID conversion. This
will be enabled later instead.