Improve efficiency of data at-rest
Wrapping the data in a loadstring'd function prevents the interpreter
from putting all the string pieces of the tables into the constant table
of the script, thereby lowering memory usage until the data is actively
read.

This brings the memory usage on retail at-rest to ~700 KB, peaking
at 4 MB when all the databases are loaded.