Remove lock-step logic from GetSoundKits
We'll be using databases from different versions of the game for
Retail going forward as soundkitname no longer exists in patch 8.3.

As a result, we'll now collect all kits into a table, add the names
for the ones present in the database, and then prune those which
had no names present.