Add zonemusic and soundkit override support
The zonemusic database will be scanned as a fallback to find names
for soundkits. Names obtained from this database will only be
applied if the soundkit doesn't have an explicit name inside the
soundkitname database.

Soundkits may be overridden in config files to forcefully include
or exclude them based on ID, or to assign custom names. This allows
us an escape hatch to add soundkits in if they're not detected due
to a lack of name or improper typing.

Updated some documentation elsewhere to unify the use of "soundkit"
instead of "sound kit".