Change load behaviour to always pull in embeds
The documented behaviour for embedded dependencies was that the
addon pulling in the library should manually load them so that
our dependencies can be omitted in packaged builds.

However this can prove somewhat inconvenient and a bit of a trap
since the "expected" behaviour is that pulling in the library
should generally just work.