Backport branched changes for SQL-based export
This ports the changes made in the "1.2" branch around using a
monolithic SQL script to do all the data collection and ties it into the
existing data storage format used by the current version of the library.

With this work changes have also been made to migrate to wago.tools for
database exports so that things are working once more.

Note that at-present Wrath data cannot be regenerated as upstream DBDs
haven't associated names to soundkit table columns; as such the changes
in that file were manual and aren't actually tested.