LibRPMedia v1.2
This changeset significantly overhauls the library interface,
deprecating the old "1.0" major version and migrating to a new
more streamlined interface with better support for auxilliary tasks
that are common in RP adodns.

The "1.0" interface is kept around for compatibility, however will
only forward calls to the new "1.2" interface.

The storage of data has been simplified and is no longer decompressed
at runtime - as such static memory usage will be higher, however
this simplifies the maintenance of the library and at this point the
gains of the compression have been outweighed by the explosive growth
of media over the past four or so years.

The export of data has been simplified and is now largely handled by
one monolithic SQL script with a few Lua function bindings. The export
process is now also automated.